+++ app/api/controllers/project/handle_get.go
@@ -14,7 +14,7 @@ func (s *ProjectController) Get(c *gin.Context) {
 		utils.Error(c, http.StatusBadRequest, err)
 		return
 	}
-	project, err := s.projectService.Get(c.Request.Context(), uint(id))
+	project, err := s.projectService.Get(c.Request.Context(), id)
 	if err != nil {
 		utils.Error(c, http.StatusBadRequest, err)
 		return
