+++ app/models/model_board.go
@@ -9,10 +9,10 @@ import (
 )
 
 type Board struct {
-	ID    int64  `gorm:"primaryKey" json:"id"`
+	ID    int    `gorm:"primaryKey" json:"id"`
 	Title string `gorm:"not null" json:"title"`
 
-	ProjectID uint `gorm:"not null;index" json:"project_id"`
+	ProjectID int `gorm:"not null;index" json:"project_id"`
 
 	CreatedAt time.Time  `json:"created_at"`
 	UpdatedAt *time.Time `json:"updated_at"`
